# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: S-PAY API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Permanently delete the current account and its data (required for App Store / Play Store compliance)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/oauth/google:
    post:
      operationId: googleTokenSignIn
      tags: [auth]
      summary: Sign in / sign up with a Google ID token (native mobile flow)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleTokenRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/oauth/apple:
    post:
      operationId: appleTokenSignIn
      tags: [auth]
      summary: Sign in / sign up with Apple (native iOS flow)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleTokenRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        phoneNumber:
          type: string
        country:
          type: string
          maxLength: 56
        accountType:
          type: string
          enum: [personal, business]
          default: personal
          description: "personal → Noah KYC · business → Noah KYB (business + representative verification)"
        businessName:
          type: string
          maxLength: 120
          description: Required when accountType is business — goes on the business virtual accounts
        signupSource:
          type: string
          maxLength: 64
          description: "Where the signup came from, e.g. jobs, jobs:<jobId>, landing, mobile"
      required:
        - email
        - password
        - fullName
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    GoogleTokenRequest:
      type: object
      properties:
        idToken:
          type: string
          description: Google ID token obtained on-device (Android / iOS)
        signupSource:
          type: string
          maxLength: 64
      required:
        - idToken
    AppleTokenRequest:
      type: object
      properties:
        identityToken:
          type: string
          description: Apple identity token from Sign in with Apple
        fullName:
          type: string
          description: Provided by Apple only on first authorization
        signupSource:
          type: string
          maxLength: 64
      required:
        - identityToken
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
        kycStatus:
          type: string
          enum: [pending, approved, rejected]
        isAdmin:
          type: boolean
        avatarUrl:
          type: string
        celoWalletAddress:
          type: string
          nullable: true
          description: "EVM wallet address on the Celo network — provisioned lazily by the active wallet provider on the user's first money action (never at signup/login, so jobs-only users cost zero WaaS MAUs)"
        accountType:
          type: string
          enum: [personal, business]
        businessName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - fullName
        - kycStatus
        - isAdmin
        - createdAt
...